+++ WebStore/Views/Employees/Details.cshtml
@@ -1,0 +1,19 @@
+﻿@model Employee
+
+
+@{
+    ViewData["Title"] = "Сотрудник № " + Model.Id;
+}
+
+<div class="container">
+    <h3>@ViewData["Title"]</h3>
+    <hr />
+    <p>@Model.SurName</p>
+    <p>@Model.FirstName</p>
+    <p>@Model.Patronymic</p>
+    <p>@Model.Age</p>
+    
+    <hr />
+    <a asp-controller="Employees" asp-action="Index">Сотрудники</a>
+</div>
+
