--- WebStore/Views/Home/Employee.cshtml
@@ -1,8 +1,0 @@
-﻿@model WebStore.Models.Employee
-
-<div class="container">
-    <h3>Сотрудник ID @Model.Id</h3>
-    <p>@Model.FirstName</p>
-    <hr />
-    <a asp-controller="Home" asp-action="Employees">Сотрудники</a>
-</div>
