+++ WebStore/Views/Home/Index.cshtml
@@ -9,7 +9,7 @@
             </div>
             <div class="col-sm-9 padding-right">
 
-                <partial name="Partial/_FeaturesItems" />
+                <partial name="Partial/_FeaturesItems" /> 
 
                 <partial name="Partial/_CategoryTab" />
